Authorization

+++ app/Http/Controllers/JobController.php
@@ -2,10 +2,11 @@
 
 namespace App\Http\Controllers;
 
-use App\Http\Controllers\Controller;
 use App\Models\Job;
+use App\Models\User;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 
 class JobController extends Controller
 {
@@ -46,16 +47,12 @@ class JobController extends Controller
 
     public function edit(Job $job)
     {
-        if (Auth::guest()) {
-            return redirect('/login');
-        }
-
         return view('jobs.edit', ['job' => $job]);
     }
 
     public function update(Job $job)
     {
-        // authorize (On hold...)
+        Gate::authorize('edit-job', $job);
 
         request()->validate([
             'title' => ['required', 'min:3'],
@@ -72,7 +69,7 @@ class JobController extends Controller
 
     public function destroy(Job $job)
     {
-        // authorize (On hold...)
+        Gate::authorize('edit-job', $job);
 
         $job->delete();
